==> Code/2. Naive Bayes Model.ipynb <==
#Load file from drive
from google.colab import drive
drive.mount('/content/drive')

# Import the necessary libraries

---



import pandas as pd
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.model_selection import train_test_split
from sklearn.naive_bayes import MultinomialNB
from sklearn.metrics import accuracy_score, precision_score, recall_score, f1_score

# Load and process initial data

# Load the dataset
file_path = '/content/drive/My Drive/Thesis/data.xlsx'
data = pd.read_excel(file_path)

# Preprocess the data by dropping unnecessary columns
data = data[['name', 'gender']]

# Convert the names to strings if not already
data['name'] = data['name'].astype(str)

# Gender label encoding

# Encode the gender labels
data['gender_encoded'] = data['gender'].map({'M': 0, 'F': 1})

# Model training

# Split the data into training and testing sets
X_train, X_test, y_train, y_test = train_test_split(
    data['name'],
    data['gender_encoded'],
    test_size=0.2,
    random_state=42
)

# Vectorize the names
vectorizer = CountVectorizer()
X_train_vect = vectorizer.fit_transform(X_train)
X_test_vect = vectorizer.transform(X_test)

# Train the MultinomialNB classifier
classifier = MultinomialNB()
classifier.fit(X_train_vect, y_train)

# Predict on the test set
y_pred = classifier.predict(X_test_vect)

# Evaluation

# Calculate accuracy, precision, recall, f1
accuracy = accuracy_score(y_test, y_pred)
precision = precision_score(y_test, y_pred, zero_division=0)
recall = recall_score(y_test, y_pred, zero_division=0)
f1 = f1_score(y_test, y_pred, zero_division=0)
print(f'Accuracy: {accuracy * 100:.2f}%')
print(f'Precision: {precision * 100:.2f}%')
print(f'Recall: {recall * 100:.2f}%')
print(f'F1 Score: {f1 * 100:.2f}%')
