==> Code/3. ANNs Model.ipynb <==
#Load file from drive
from google.colab import drive
drive.mount('/content/drive')

#!pip install pandas numpy scikit-learn tensorflow keras imbalanced-learn

# Import the necessary libraries

import pandas as pd
import numpy as np
import pickle
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import LabelEncoder
from sklearn.metrics import precision_score, recall_score, f1_score
from imblearn.over_sampling import SMOTE
from tensorflow.keras.preprocessing.text import Tokenizer
from tensorflow.keras.preprocessing.sequence import pad_sequences
from tensorflow.keras.models import Sequential
from tensorflow.keras.layers import Embedding, Bidirectional, LSTM, Dense
from tensorflow.keras.utils import to_categorical
from tensorflow.keras.metrics import Precision


# Load and process initial data

# Load the dataset
file_path = '/content/drive/My Drive/Thesis/data.xlsx'
data = pd.read_excel(file_path)

# Preprocess the data by dropping unnecessary columns
data = data[['name', 'gender']]

# Convert the names to strings if not already
data['name'] = data['name'].astype(str)

# Gender label encoding

label_encoder = LabelEncoder()
data['gender_encoded'] = label_encoder.fit_transform(data['gender'])

# Save label encoder
with open('/content/drive/My Drive/Thesis/model/label_encoder.pickle', 'wb') as file:
    pickle.dump(label_encoder, file, protocol=pickle.HIGHEST_PROTOCOL)

# Tokenize names (char-level) and handle length

# Tokenize the names at the character level
tokenizer = Tokenizer(char_level=True)
tokenizer.fit_on_texts(data['name'])
sequences = tokenizer.texts_to_sequences(data['name'])
max_length = max([len(seq) for seq in sequences])
vocab_size = len(tokenizer.word_index) + 1
data_sequences = pad_sequences(sequences, maxlen=max_length, padding='post')

# Save tokenizer
with open('/content/drive/My Drive/Thesis/model/tokenizer.pickle', 'wb') as handle:
    pickle.dump(tokenizer, handle, protocol=pickle.HIGHEST_PROTOCOL)

# Split train/test data and handle imbalance using SMOTE

# Split the data
X_train, X_test, y_train, y_test = train_test_split(data_sequences, data['gender_encoded'],
                                                    test_size=0.2, random_state=42)

# Apply SMOTE
smote = SMOTE(random_state=42)
X_train_smote, y_train_smote = smote.fit_resample(X_train, y_train)

# BLSTM model definition and traning model

def create_blstm_model(input_dim, output_dim, input_length):
    model = Sequential()
    model.add(Embedding(input_dim=input_dim, output_dim=output_dim, input_length=input_length))
    model.add(Bidirectional(LSTM(64)))
    model.add(Dense(64, activation='relu'))
    model.add(Dense(2, activation='softmax'))
    model.compile(optimizer='adam', loss='categorical_crossentropy', metrics=['accuracy', Precision()])
    return model

print("\nTraining and evaluating BLSTM model...")

# Create and compile the BLSTM model
model = create_blstm_model(vocab_size, 50, max_length)

# One-hot encode labels
y_train_encoded = to_categorical(y_train_smote)
y_test_encoded = to_categorical(y_test)

# Train the model
model.fit(X_train_smote, y_train_encoded,
          batch_size=32, epochs=10,
          validation_data=(X_test, y_test_encoded))

# Evaluate the model on the test set
loss, accuracy, precision = model.evaluate(X_test, y_test_encoded, verbose=0)

# Forecast
y_pred = model.predict(X_test, verbose=0)
y_pred_labels = np.argmax(y_pred, axis=1)
y_test_labels = np.argmax(y_test_encoded, axis=1)

# Calculate evaluation indexes
precision = precision_score(y_test_labels, y_pred_labels, average='weighted')
recall = recall_score(y_test_labels, y_pred_labels, average='weighted')
f1 = f1_score(y_test_labels, y_pred_labels, average='weighted')

# Print the results
print(f'Accuracy: {accuracy * 100:.2f}%')
print(f'Precision: {precision * 100:.2f}%')
print(f'Recall: {recall * 100:.2f}%')
print(f'F1 Score: {f1 * 100:.2f}%')

# Save model
model.save('/content/drive/My Drive/Thesis/model/blstm_gender_model.h5')
